requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1

# OpenAI
openai>=1.30.0

# Supabase (latest version)
supabase>=2.9.0

# Additional dependencies
html5lib==1.1
lxml==5.1.0
